requirements.txt
django
djangorestframework
django-cors-headers
pytest
pytest-django
factory_boy
faker
black